% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CookDisLasso.R
\name{CookDisLasso}
\alias{CookDisLasso}
\title{Calculate case influence for the lasso and plot Case influence graph}
\usage{
CookDisLasso(
  X,
  y,
  k,
  fineness,
  s,
  mode = c("fraction", "norm", "lambda"),
  threshold = TRUE,
  plot = TRUE
)
}
\arguments{
\item{X}{matrix of predictors}

\item{y}{response}

\item{k}{an integer, or vector of integers representing the index/indices
of observations that are of interest. By default, CookDisLasso considers
all observations.}

\item{fineness}{number of grid points for fractions being considered from 0 to 1.
The default is set to be 100. Then we take fractions 0, 0.01, ..., 0.99}

\item{s}{a value, or vector of values, indexing the penalty levels to consider.
Its values depends on the mode= argument. By default (mode="fraction"),
s should take on values between 0 and 1.}

\item{mode}{Mode="fraction", then s should be a number between 0 and 1, and it refers
to the ratio of the L1 norm of the coefficient vector, relative to the
norm at the full LS solution if n>p or the minimum L1 norm linear
interpolation solution if n<=p. Mode="norm" means s refers to the L1 norm
of the coefficient vector. Mode="lambda" uses the lasso regularization
parameter for s. Abbreviations allowed.}

\item{threshold}{If TRUE, CookDisLasso prints the threshold for influential points. This
can only set to be TRUE if k is 1:n (by default).}

\item{plot}{If TRUE, CookDisLasso plots the case influence graph}
}
\value{
CD_Mat          matrix n by f      matrix of cook's distance at each fraction \cr
Lambda_list     vector f by 1      a vector of lambdas at each fraction \cr
fraction        vector f by 1      a vector of fractions based on required fineness \cr
threshold_table matrix f by 2      threshold at each fraction \cr
beta_hat_table  matrix p by f      a matrix of betahats at each fraction \cr
fig        the case influence graph \cr
where f means the number of fractions CookDisLasso considers. Its either |k| or fineness.
}
\description{
Using case-weight lasso model and the solution path algorithm, this function calculates
the case influence for the lasso for all observations(for specified subset), all fractions(for specified lambdas) and also gives
case influence graph.
}
\details{
\enumerate{
\item threshold needs case influence of all observations to calculate. So threshold should be set to FALSE if the functions only calculate
case influence for a subset of observations,i.e. k is not NULL.
\item fineness automatically chooses a list of lambda candidates from 0 to maximum lambda that penalizes all beta to 0. If lambda candidate
is pre-specified, i.e. lambda is not NULL, fineness should be set to NULL.
}
}
\examples{
library(lars)
data(diabetes)
attach(diabetes)
CookDisLasso(x,y)
detach(diabetes)

}
