% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CookDisLasso.R
\name{plot.CookDisLasso}
\alias{plot.CookDisLasso}
\title{print a CookDisLasso object}
\usage{
\method{plot}{CookDisLasso}(
  object,
  mode = c("case-influence-graph", "residual-leverage-plot"),
  which.one,
  ...
)
}
\arguments{
\item{object}{fitted CookDisLasso object}

\item{mode}{specify which graph to print: 'case-influence-graph' or 'residual-leverage-plot'}

\item{which.one}{a integer specifying which fraction to print if mode=='residual-leverage-plot'}
}
\value{
the required plot
}
\description{
this function is built for leverage-residual plot and case influence graph. \cr
\itemize{
\item Leverage-residual plot records the leverage and residual of obs and their Lasso case influence at one given fraction
\item Case influence graph records the Lasso case influence of obs as fraction goes from 1 to 0
}
}
\details{
'which.one' should be an integer indexing the fraction of interest.
The fraction of interest should be "object$fraction(which.one)".
If mode is set to be case-influence-graph, the function outputs case influence under all fractions.
}
\examples{
library(lars)
data(diabetes)
attach(diabetes)
CookDisLasso(x,y)
detach(diabetes)

set.seed(10)
x = matrix(rnorm(200*50),nrow=200)
y = x[,1:5]\%*\%c(5,4,3,2,1) + rnorm(200)
obj1 = CookDisLasso(x,y, k=1:10, s=c(0.1,0.2), mode = "fraction", threshold = FALSE)
plot(obj1, 'resid', 1)
obj2 = CookDisLasso(x,y, fineness=40, threshold = TRUE)
plot(obj2, 'case-influence-graph')
}
