% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolPathLasso.R
\name{plot.SolPathLooLasso}
\alias{plot.SolPathLooLasso}
\title{plot the solution path from a SolPathLooLasso object}
\usage{
\method{plot}{SolPathLooLasso}(object, plot = 2, ...)
}
\arguments{
\item{object}{fitted SolPathLooLasso object}

\item{plot}{If plot=1, this function plots the linear approximated solution path.
If plot=2, plot exact solution path.}
}
\value{
The solution path plot
}
\description{
plot the solution path from a SolPathLooLasso object
}
\examples{
library(lars)
data(diabetes)
attach(diabetes)
obj = SolPathLooLasso(x,y,k = 182, 5, 'lambda')

detach(diabetes)

set.seed(100)
x = matrix(rnorm(50*200),nrow=50)
y = x[,1:5]\%*\%c(5,4,3,2,1) + rnorm(50)
ans = SolPathLooLasso(x,y,k = 1, s = 0.7, mode = "fraction")
plot(ans)
predict(ans)
coef(ans)
}
\seealso{
predict, coef methods
}
