% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolPathLasso.R
\name{predict.SolPathLooLasso}
\alias{predict.SolPathLooLasso}
\title{extract coefficients or estimates from a SolPathLooLasso object}
\usage{
\method{predict}{SolPathLooLasso}(obj, newx, type = c("fit", "coefficients"), ...)
}
\arguments{
\item{obj}{fitted SolPathLooLasso object}

\item{newx}{Matrix of new values for x at which predictions are to be made.
Must be a matrix.
This argument is not used for type="coefficients"}

\item{type}{Type of prediction required.
Type "coefficients" computes the LOO Lasso estimate
Type "fit" computes the y estimates of given newx}
}
\value{
The object returned depends on type.
}
\description{
extract coefficients or estimates from a SolPathLooLasso object
}
\examples{
library(lars)
data(diabetes)
attach(diabetes)
obj = SolPathLooLasso(x,y,k = 182, 5, 'lambda')

detach(diabetes)

set.seed(100)
x = matrix(rnorm(50*200),nrow=50)
y = x[,1:5]\%*\%c(5,4,3,2,1) + rnorm(50)
ans = SolPathLooLasso(x,y,k = 1, s = 0.7, mode = "fraction")
plot(ans)
predict(ans)
coef(ans)
}
\seealso{
plot, coef methods
}
