% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidLevPlot.R
\name{ResidLevPlot}
\alias{ResidLevPlot}
\title{Plot the residuals and leverages of all observations against their case influence when penalty is fixed}
\usage{
ResidLevPlot(X, y, lambda, studentize = TRUE)
}
\arguments{
\item{X}{matrix n by p      design matrix}

\item{y}{vector n by 1      response vector}

\item{lambda}{number or vector   if only one or a subset of penalty parameters are of interest}

\item{studentize}{Boolean            whether to studentize the residual}
}
\value{
A plot includes all observations. X axis: leverage(when w=1), Y axis: residuals
}
\description{
Each point in the plot represents a observation and its size indicates its case influence for the lasso.
Observations that are identified as influential points (having case influence greater than the threshold) is marked in red.
}
\examples{
library(lars)
data(diabetes)
attach(diabetes)
ResidLevPlot(x,y,1, FALSE)
detach(diabetes)

x = matrix(rnorm(50*200),nrow=50)
y = x[,1:5]\%*\%c(5,4,3,2,1) + rnorm(50)
ResidLevPlot(x,y,3)
}
