% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolPathLasso.R
\name{SolPathLasso}
\alias{SolPathLasso}
\title{Output the solution path of observation k when lambda is fixed.
Only covariates that have sign change are included in the solution path plot}
\usage{
SolPathLasso(X, y, k = 1, lambda = 50, plot = 0, lb = 0)
}
\arguments{
\item{X}{matrix n by p      design matrix}

\item{y}{vector n by 1      response vector}

\item{k}{integer            observation of interest}

\item{lambda}{float > 0          penalty parameter}

\item{plot}{0                  no plot
1                  plot approximate
2                  plot exact solution path}

\item{lb}{float < 1          the lower bound of omega}
}
\value{
w_path          vector b by 1      a vector of breakpoints
hkk_path        vector b by 1      leverages at each breakpoint
beta_path       matrix b by p      beta estimate at each breakpoint
s_path          matrix b by p      beta hat's sign at each breakpoint
beta0_path      vector b by 1      beta0 at each breakpoint
l1norm          vector b by 1      l1 norm of beta at each breakpoint
the solution path visualization
}
\description{
Output the solution path of observation k when lambda is fixed.
Only covariates that have sign change are included in the solution path plot
}
\examples{
library(lars)
data(diabetes)
attach(diabetes)
SolPathLasso(x,y,k = 182, lambda = 5, plot = 2)
detach(diabetes)

set.seed(100)
}
