% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CookDisLasso.R
\name{CookDisLasso}
\alias{CookDisLasso}
\title{calculate case influence for Lasso regression}
\usage{
CookDisLasso(
  X,
  y,
  k,
  fineness,
  s,
  mode = c("fraction", "norm", "lambda"),
  threshold = TRUE
)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation vector.
Requirement: nvars >1; in other words, x should have 2 or more columns.}

\item{y}{response variable.}

\item{k}{an integer, or vector of integers representing the index/indices
of observations that are of interest. By default, \code{CookDisLasso()} considers
all observations.}

\item{fineness}{number of grid points for fractions being considered from 0 to 1.
The default is set to be 100. Then we take fractions 0, 0.01, ..., 0.99}

\item{s}{a value, or vector of values, indexing the penalty levels to consider.
Its values depends on the mode= argument. By default (mode="fraction"),
s should take on values between 0 and 1.}

\item{mode}{Mode="fraction", then s should be a number between 0 and 1, and it refers
to the ratio of the L1 norm of the coefficient vector, relative to the
norm at the full LS solution if n>p or the minimum L1 norm linear
interpolation solution if n<=p. Mode="norm" means s refers to the L1 norm
of the coefficient vector. Mode="lambda" uses the lasso regularization
parameter for s. Abbreviations allowed.}

\item{threshold}{If TRUE, CookDisLasso prints the threshold for influential points. This
can only set to be TRUE if k is 1:n (by default).}
}
\value{
\item{CD_Mat}{matrix of cook's distance at each fraction}
\item{Lambda_list}{a vector of lambdas at each fraction}
\item{fraction}{a vector of fractions based on required fineness}
\item{threshold_table}{threshold at each fraction if threshold = TRUE}
\item{beta_hat_table}{a matrix of betahats at each fraction}
\item{index}{samples of interest. If k is specified in input, index = k, otherwise 1:n}
\item{lev_history}{record the leverages of samples of interest in each grid point specified in fineness}
\item{denom}{estimated error variance}
}
\description{
Using case-weight lasso model and the solution path algorithm, this function calculates
the case influence for the lasso for all observations(for specified subset), all fractions(for specified lambdas) and also gives
case influence graph.
}
\details{
\enumerate{
\item threshold needs case influence of all observations to calculate. So threshold should be set to FALSE if the functions only calculate
case influence for a subset of observations,i.e. k is not NULL.
\item fineness automatically chooses a list of lambda candidates from 0 to maximum lambda that penalizes all beta to 0. If lambda candidate
is pre-specified, i.e. lambda is not NULL, fineness should be set to NULL.
}
}
\examples{
library(lars)
data(diabetes)
attach(diabetes)
CookDisLasso(x,y)
detach(diabetes)

set.seed(10)
x = matrix(rnorm(200*50),nrow=200)
y = x[,1:5]\%*\%c(5,4,3,2,1) + rnorm(200)
obj1 = CookDisLasso(x,y, k=1:10, s=c(0.1,0.2), mode = "fraction", threshold = FALSE)
plot(obj1, 'resid', 1)
obj2 = CookDisLasso(x,y, fineness=40, threshold = TRUE)
plot(obj2, 'case-influence-graph')
}
\seealso{
plot method
}
